Add ElasticSearchController
Automate one test suite with the controller.

Currently ElasticIndexingStorageTest.testPublic fails with the error:

java.io.IOException: POST
http://localhost:35787/test_1501273022012.publishable/data/_update_by_query:
HTTP/1.1 500 Internal Server Error
{"error":{"root_cause":[{"type":"circuit_breaking_exception","reason":"[script]
Too many dynamic script compilations within one minute, max: [15/min];
please use on-disk, indexed, or scripts with parameters instead; this
limit can be changed by the [script.max_compilations_per_minute]
setting","bytes_wanted":0,"bytes_limit":0}],"type":"general_script_exception","reason":"Failed
to compile inline script [ctx._source.public = false;] using lang
[painless]","caused_by":{"type":"circuit_breaking_exception","reason":"[script]
Too many dynamic script compilations within one minute, max: [15/min];
please use on-disk, indexed, or scripts with parameters instead; this
limit can be changed by the [script.max_compilations_per_minute]
setting","bytes_wanted":0,"bytes_limit":0}},"status":500}

Need to investigate further. The other 4 tests pass.

+++ test/src/kbaserelationengine/search/test/ElasticIndexingStorageTest.java
@@ -2,6 +2,8 @@ package kbaserelationengine.search.test;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashSet;
@@ -42,26 +44,27 @@ import kbaserelationengine.search.ObjectData;
 import kbaserelationengine.search.PostProcessing;
 import kbaserelationengine.system.IndexingRules;
 import kbaserelationengine.system.ObjectTypeParsingRules;
+import kbaserelationengine.test.common.TestCommon;
+import kbaserelationengine.test.controllers.elasticsearch.ElasticSearchController;
 import us.kbase.common.service.UObject;
 
 public class ElasticIndexingStorageTest {
     
-    //TODO NOW Automate tests es
-    
-    public static final boolean cleanup = true;
-    
     private static ElasticIndexingStorage indexStorage;
     private static File tempDir = null;
     private static ObjectLookupProvider objLookup;
+    private static ElasticSearchController es;
     
     @BeforeClass
     public static void prepare() throws Exception {
-        tempDir = new File("test_local/temp_files/esbulk");
+        final Path tdir = Paths.get(TestCommon.getTempDir());
+        tempDir = tdir.resolve("ElasticIndexingStorageTest").toFile();
+        FileUtils.deleteQuietly(tempDir);
+        es = new ElasticSearchController(TestCommon.getElasticSearchExe(), tdir);
         String indexNamePrefix = "test_" + System.currentTimeMillis() + ".";
         indexStorage = new ElasticIndexingStorage(
-                new HttpHost("localhost", 9200), tempDir);
+                new HttpHost("localhost", es.getServerPort()), tempDir);
         indexStorage.setIndexNamePrefix(indexNamePrefix);
-        cleanup();
         tempDir.mkdirs();
         objLookup = new ObjectLookupProvider() {
             
@@ -120,22 +123,10 @@ public class ElasticIndexingStorageTest {
     
     @AfterClass
     public static void teardown() throws Exception {
-        if (cleanup) {
-            cleanup();
+        if (es != null) {
+            es.destroy(TestCommon.getDeleteTempFiles());
         }
-    }
-    
-    private static void cleanup() throws Exception {
-        Set<String> indNames = indexStorage.listIndeces();
-        for (String index : indNames) {
-            if (!index.startsWith("test_")) {
-                System.out.println("Skipping Elasticsearch index: " + index);
-                continue;
-            }
-            System.out.println("Deleting Elasticsearch index: " + index);
-            indexStorage.deleteIndex(index);
-        }
-        if (tempDir != null && tempDir.exists()) {
+        if (tempDir != null && tempDir.exists() && TestCommon.getDeleteTempFiles()) {
             FileUtils.deleteQuietly(tempDir);
         }
     }
